use correct types
