+++ frontend/src/metabase-types/api/group.ts
@@ -1,0 +1,7 @@
+export type GroupId = number;
+
+export type Group = {
+  id: GroupId;
+  name: string;
+  member_count: number;
+};
