+++ frontend/src/metabase-types/api/index.ts
@@ -2,7 +2,10 @@ export * from "./automagic-dashboards";
 export * from "./collection";
 export * from "./dashboard";
 export * from "./database";
+export * from "./table";
 export * from "./timeline";
 export * from "./settings";
 export * from "./slack";
 export * from "./user";
+export * from "./group";
+export * from "./permissions";
