+++ frontend/src/metabase-types/api/table.ts
@@ -1,0 +1,6 @@
+export interface Table {
+  id: number;
+  db_id: number;
+  name: string;
+  schema: string;
+}
